requirements.txt
unbabel-comet
ollama
matplotlib
seaborn
scikit-learn